Fix build error and CodeQL finding
- weblog.c: rename LINE_MAX -> WEBLOG_LINE_MAX. It collided with
  picolibc's sys/syslimits.h LINE_MAX (pulled in transitively via
  FreeRTOS.h), which -Werror turned into a hard build failure
  ('LINE_MAX redefined').
- http_server.c: cast esp_timer_get_time()/1000000 to (long long)
  before the %lld in status_handler's snprintf. CodeQL's default C/C++
  setup runs with build-mode:none (no real ESP-IDF headers), so it
  couldn't resolve esp_timer_get_time()'s actual int64_t return type
  and flagged a false type mismatch; the explicit cast removes the
  ambiguity for any analyzer.

+++ src/http_server.c
@@ -105,7 +105,7 @@ static esp_err_t status_handler(httpd_req_t *req)
              "\"ip\":\"%s\",\"rssi\":%d,\"retries\":%d},"
              "\"ap\":{\"ssid\":\"PowerStream-Bridge\",\"ip\":\"192.168.4.1\","
              "\"clients\":%d}}",
-             esp_timer_get_time() / 1000000,
+             (long long)(esp_timer_get_time() / 1000000),
              (unsigned long)esp_get_free_heap_size(),
              w.sta_enabled ? "true" : "false",
              w.sta_connected ? "true" : "false",
